lpddr5: commands: Reverse the order of RD / WR cmds and WCK sync
Although the logic is combinational and it shouldn't matter in what order
the wires are assigned, the previous approach of first translating into
the commands and then issuing the WCK sync would cause the CAS to be send
without WS_RD / WS_WR parameter (and therefore issuing an illegal command).
Moreover, this issue was only observed on hardware with no reported
timing violations.
The simphy / s7phy simulations followed expected behaviour.

Internal-tag: [#55746]
Signed-off-by: Wiktoria Kuna <wkuna@antmicro.com>